src/requirements.txt
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-dialogs==4.17.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
azure-identity>=1.15.0
requests>=2.28.0
botframework-connector>=4.17.0
 